backed refactored

+++ Backend/Controller/AdminController.js
@@ -2,9 +2,12 @@ import jwt from "jsonwebtoken";
 import UserModel from "../Model/UserModel.js";
 import TurfModel from "../Model/TurfModel.js";
 import bookingModel from "../Model/BookingModel.js";
+
+//admin credentials
+const adminEmail = process.env.ADMIN_EMAIL;
+const adminPassword = process.env.ADMIN_PASSWORD;
+
 export const adminLogin = async (req, res) => {
-  const adminEmail = process.env.ADMIN_EMAIL;
-  const adminPassword = process.env.ADMIN_PASSWORD;
   const { email, password } = req.body;
   try {
     if (email === adminEmail && password === adminPassword) {
@@ -18,10 +21,13 @@ export const adminLogin = async (req, res) => {
 
 export const getRequestedTurf = async (req, res) => {
   try {
-    const turfs = await TurfModel.find({ turfStatus: false });
-    if (turfs.length === 0)
-      return res.status(404).json({ message: "No turfs found" });
-    res.status(200).json(turfs);
+    const admin = req.user.id;
+    if (adminEmail === admin) {
+      const turfs = await TurfModel.find({ turfStatus: false });
+      if (turfs.length === 0)
+        return res.status(404).json({ message: "No turfs found" });
+      res.status(200).json(turfs);
+    }
   } catch (error) {
     console.log(error);
     res.status(500).json({ message: error.message });
@@ -29,15 +35,16 @@ export const getRequestedTurf = async (req, res) => {
 };
 
 export const toAcceptTurfRequest = async (req, res) => {
+  const admin = req?.user?.id;
   try {
-    console.log(req.body);
-    const ID = req.body.ID;
-    console.log(ID);
-    const turf = await TurfModel.findByIdAndUpdate(ID, {
-      turfStatus: true,
-    });
-    console.log(turf);
-    res.status(200).json(turf.turfStatus);
+    if (adminEmail === admin) {
+      const ID = req.body.ID;
+      const turf = await TurfModel.findByIdAndUpdate(ID, {
+        turfStatus: true,
+      });
+      console.log(turf);
+      res.status(200).json(turf.turfStatus);
+    }
   } catch (error) {
     console.log(error);
     res.status(500).json({ message: error.message });
@@ -45,17 +52,27 @@ export const toAcceptTurfRequest = async (req, res) => {
 };
 
 export const toCancelTurfRequest = async (req, res) => {
-  const ID = req.body.id;
-  await TurfModel.findByIdAndDelete(ID);
-  res.status(200).json({ message: "cancelled" });
+  const admin = req.user.id;
+  try {
+    if (adminEmail === admin) {
+      const ID = req.body.id;
+      await TurfModel.findByIdAndDelete(ID);
+      res.status(200).json({ message: "cancelled" });
+    }
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
 };
 
 export const toViewUserList = async (req, res) => {
+  const admin = req.user.id;
   try {
-    const Users = await UserModel.find();
-    if (Users.length === 0)
-      return res.status(404).json({ message: "No users found" });
-    return res.status(200).json(Users);
+    if (adminEmail === admin) {
+      const Users = await UserModel.find();
+      if (Users.length === 0)
+        return res.status(404).json({ message: "No users found" });
+      return res.status(200).json(Users);
+    }
   } catch (error) {
     console.log(error);
     res.status(500).json({ message: error.message });
@@ -63,11 +80,14 @@ export const toViewUserList = async (req, res) => {
 };
 
 export const toGetAllTurfs = async (req, res) => {
+  const admin = req.user.id;
   try {
-    const turfs = await TurfModel.find({ turfStatus: true });
-    if (turfs.length === 0)
-      return res.status(404).json({ message: "Turfs not found" });
-    res.status(200).json(turfs);
+    if (adminEmail === admin) {
+      const turfs = await TurfModel.find({ turfStatus: true });
+      if (turfs.length === 0)
+        return res.status(404).json({ message: "Turfs not found" });
+      res.status(200).json(turfs);
+    }
   } catch (error) {
     console.log(error);
     res.status(500).json({ message: error.message });
@@ -75,28 +95,31 @@ export const toGetAllTurfs = async (req, res) => {
 };
 
 export const toViewReports = async (req, res) => {
+  const admin = req.user.id;
   try {
-    const report = await bookingModel.aggregate([
-      { $match: { payment: "Success" } },
-      {
-        $lookup: {
-          from: "turfs",
-          localField: "turf",
-          foreignField: "_id",
-          as: "turf",
+    if (adminEmail === admin) {
+      const report = await bookingModel.aggregate([
+        { $match: { payment: "Success" } },
+        {
+          $lookup: {
+            from: "turfs",
+            localField: "turf",
+            foreignField: "_id",
+            as: "turf",
+          },
         },
-      },
-      {
-        $group: {
-          _id: "$turf._id",
-          name: { $first: "$turf.turfName" },
-          totalPrice: { $sum: "$price" },
-          count: { $sum: 1 },
+        {
+          $group: {
+            _id: "$turf._id",
+            name: { $first: "$turf.turfName" },
+            totalPrice: { $sum: "$price" },
+            count: { $sum: 1 },
+          },
         },
-      },
-      { $sort: { totalPrice: 1 } },
-    ]);
-    res.status(200).json(report)
+        { $sort: { totalPrice: 1 } },
+      ]);
+      res.status(200).json(report);
+    }
   } catch (error) {
     console.log(error);
     res.status(500).json({ message: error.message });
